refactor(renderer): decompose Sidebar into a scope plus zero-prop components
`Sidebar()` was one 1930-line function holding 31 reactive primitives, 17
imperative drag locals, five JSX-returning closures and four regions spliced
together by one drag&drop system. The pure logic had already been extracted
(`sidebar-sections.ts`, `sidebar-pins.ts`, `sidebar-people-order.ts`,
`sidebar-agent-states.ts`) but the view was never broken up, so every change
still meant reading the whole file.

This finishes the pattern `conversation/` already established: the parent takes
props, builds a scope, publishes it through context, and the children are
zero-prop.

- `Sidebar.tsx` 2338 -> 80 lines. `SidebarProps` is unchanged except for
  `export`; `WorkspaceSidebar.tsx` is untouched. `SidebarAgentState` stays
  declared here so `sidebar-agent-states.ts` and the stories keep their imports.
- `sidebar-scope.ts` holds everything the component was minus the markup;
  `createSidebarDragEngine.ts` holds the imperative half, which needs a closure
  of its own.
- Pure, node-testable modules: `sidebar-drag-model.ts`, `sidebar-drag-dom.ts`,
  `sidebar-filtering.ts`.
- 13 zero-prop components under `sidebar/`.

The four drag-offset transports are unified. One computation (`reorderOffsets`),
one transport (`drag.offsets`, written per key so a row still only re-renders
when its own offset moves), one read (`dragOffset(id)`). The index-ternary the
pinned and agent regions used and the splice-then-reindex the people and section
regions used are the same algorithm once placement comes from `placementForSwap`,
so this is an equivalence - except that people move from same-tick imperative
`style.setProperty` writes to a store write published on the next flush, which is
where sections already were.

Two reactivity bugs the decomposition surfaced and fixed: the agent context
menu's pin limit was read eagerly (as a component body it runs once, where the
closure it replaced ran inside the parent's render), and the section editor's
`onCleanup` had to move out of the ref callback, which has no owner in Solid 2.

Model: Claude Opus 5 via Claude Code.
