+++ src/renderer/src/components/sidebar/SidebarIcons.tsx
@@ -1,0 +1,57 @@
+/**
+ * The sidebar's own SVGs. `EditIcon` and `DeleteIcon` stay hand-written rather than becoming
+ * `Pencil` and `Trash2`: swapping them changes the rendered markup, so it belongs to a separate
+ * change with a before/after story - the exception `src/renderer/AGENTS.md` asks to be named.
+ */
+
+export function SearchIcon() {
+  return (
+    <svg aria-hidden="true" viewBox="0 0 20 20" class="size-4 fill-none stroke-current">
+      <circle cx="8.5" cy="8.5" r="5.5" stroke-width="1.6" />
+      <path d="m12.7 12.7 3.6 3.6" stroke-width="1.6" stroke-linecap="round" />
+    </svg>
+  );
+}
+
+export function PlusIcon() {
+  return (
+    <svg aria-hidden="true" viewBox="0 0 20 20" class="size-[18px] fill-none stroke-current">
+      <path d="M10 4v12M4 10h12" stroke-width="1.5" stroke-linecap="round" />
+    </svg>
+  );
+}
+
+export function SidebarToggleIcon() {
+  return (
+    <svg aria-hidden="true" viewBox="0 0 20 20" class="sidebar-toggle-icon">
+      <rect x="2.75" y="3.25" width="14.5" height="13.5" rx="2.25" />
+      <path d="M7.25 3.75v12.5" />
+    </svg>
+  );
+}
+
+export function EditIcon() {
+  return (
+    <svg aria-hidden="true" viewBox="0 0 20 20" class="bot-context-icon size-4 fill-none stroke-current">
+      <path d="m12.6 4.2 3.2 3.2-8.7 8.7-3.8.6.6-3.8 8.7-8.7Z" stroke-width="1.4" />
+      <path d="m10.9 5.9 3.2 3.2" stroke-width="1.4" />
+    </svg>
+  );
+}
+
+export function DeleteIcon() {
+  return (
+    <svg
+      aria-hidden="true"
+      viewBox="0 0 20 20"
+      class="bot-context-icon bot-context-danger-icon size-4 fill-none stroke-current"
+    >
+      <path
+        d="M4.5 6.2h11M8 3.8h4M6.2 6.2l.7 9.3h6.2l.7-9.3M8.4 8.7v4.5M11.6 8.7v4.5"
+        stroke-width="1.35"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+      />
+    </svg>
+  );
+}
